+++ src/rogallo/screens/user_input.py
@@ -2,7 +2,10 @@
 
 ##############################################################################
 # Python imports.
-from typing import Final
+from os import getenv
+from pathlib import Path
+from subprocess import run
+from tempfile import NamedTemporaryFile
 
 ##############################################################################
 # Textual imports.
@@ -16,6 +19,10 @@ from textual.widgets import TextArea
 ##############################################################################
 # Wasat imports.
 from wasat import GeminiURI
+
+##############################################################################
+# Local imports.
+from ..types import DEFAULT_GEMINI_EXTENSION
 
 
 ##############################################################################
@@ -48,6 +55,7 @@ class UserInput(ModalScreen[str | None]):
     BINDINGS = [
         ("escape", "escape"),
         ("f2", "submit"),
+        ("f3", "edit_externally"),
     ]
 
     _input = query_one(TextArea)
@@ -102,19 +110,22 @@ class UserInput(ModalScreen[str | None]):
         """The current query in the input area."""
         return self._location.with_query(self._current_text)
 
-    _SUBMIT: Final[str] = "Press F2 to submit"
-    """The subtitle to display when the input is valid."""
+    @property
+    def _external_editor(self) -> str | None:
+        """The external editor to use, if any."""
+        return getenv("VISUAL") or getenv("EDITOR") or None
 
     def _update_subtitle(self) -> None:
         """Update the subtitle of the input area."""
+        footer = "F2: Submit"
+        if bool(self._external_editor):
+            footer += " | F3: $EDITOR"
         if not self._input.text:
-            self._input.border_subtitle = self._SUBMIT
+            self._input.border_subtitle = footer
         elif self._current_query.is_too_long:
             self._input.border_subtitle = "Input is too long!"
         else:
-            self._input.border_subtitle = (
-                f"{self._SUBMIT} ({self._current_query.bytes_left})"
-            )
+            self._input.border_subtitle = f"{footer} ({self._current_query.bytes_left})"
 
     @on(TextArea.Changed)
     def _limit_check(self) -> None:
@@ -134,6 +145,23 @@ class UserInput(ModalScreen[str | None]):
     def action_escape(self) -> None:
         """Escape out without getting the input."""
         self.dismiss(None)
+
+    def action_edit_externally(self) -> None:
+        """Edit the input in an external editor."""
+        if not (editor := self._external_editor):
+            return
+        with NamedTemporaryFile(
+            mode="w+", delete=False, encoding="utf-8", suffix=DEFAULT_GEMINI_EXTENSION
+        ) as temp_file:
+            user_input = Path(temp_file.name)
+            temp_file.write(self._current_text)
+            temp_file.close()
+            try:
+                with self.app.suspend():
+                    run((editor, user_input))
+                self._input.text = user_input.read_text(encoding="utf-8")
+            finally:
+                user_input.unlink(missing_ok=True)
 
 
 ### user_input.py ends here