+++ ChangeLog.md
@@ -10,6 +10,8 @@
   ([#211](https://github.com/davep/rogallo/pull/211))
 - Added `!finger <user[@host]>` as a command line command.
   ([#211](https://github.com/davep/rogallo/pull/211))
+- Added support for using `$VISUAL` or `$EDITOR` to edit user input.
+  ([#212](https://github.com/davep/rogallo/pull/212))
 
 ## v1.0.0